// rs-lang/src/interfaces/IWords.ts
export interface IWords {
    id: string
    group: number
    page: number
    word: string
    image: string
    audio: string
    audioMeaning: string
    audioExample: string
    textMeaning: string
    textExample: string
    transcription: string
    wordTranslate: string
    textMeaningTranslate: string
    textExampleTranslate: string
}